Clean Up: Validation helper (#2348)
