+++ services/ui-src/src/measures/2021/globalValidations/validateOneCatRateHigherThanOtherCat/index.test.ts
@@ -11,7 +11,7 @@ import {
   higherRate,
   lowerRate,
   partialRate,
-} from "utils/testUtils/2023/validationHelpers";
+} from "utils/testUtils/validationHelpers";
 
 jest.mock("utils/getLabelText", () => ({
   isLegacyLabel: () => true,
