+++ services/ui-src/src/measures/2021/globalValidations/validatePartialRateCompletion/index.test.ts
@@ -9,7 +9,7 @@ import {
   locationDictionary,
   simpleRate,
   partialRate,
-} from "utils/testUtils/2023/validationHelpers";
+} from "utils/testUtils/validationHelpers";
 import { LabelData } from "utils";
 
 jest.mock("utils/getLabelText", () => ({
