+++ services/ui-src/src/measures/2022/shared/globalValidations/validateEqualCategoryDenominators/index.test.ts
@@ -11,7 +11,7 @@ import {
   simpleRate,
   partialRate,
   generatePmQualifierRateData,
-} from "utils/testUtils/2023/validationHelpers";
+} from "utils/testUtils/validationHelpers";
 
 jest.mock("utils/getLabelText", () => ({
   isLegacyLabel: () => true,
