+++ services/ui-src/src/measures/2022/shared/globalValidations/validateOneQualRateHigherThanOtherQual/index.test.ts
@@ -11,7 +11,7 @@ import {
   higherRate,
   lowerRate,
   partialRate,
-} from "utils/testUtils/2023/validationHelpers";
+} from "utils/testUtils/validationHelpers";
 import { LabelData } from "utils";
 
 jest.mock("utils/getLabelText", () => ({
