+++ services/ui-src/src/measures/2022/shared/globalValidations/validateRateNotZero/index.test.ts
@@ -8,7 +8,7 @@ import {
   simpleRate,
   partialRate,
   generateOtherPerformanceMeasureData,
-} from "utils/testUtils/2023/validationHelpers";
+} from "utils/testUtils/validationHelpers";
 
 jest.mock("utils/getLabelText", () => ({
   isLegacyLabel: () => true,
