+++ services/ui-src/src/measures/2022/shared/globalValidations/validateRateZero/index.test.ts
@@ -10,7 +10,7 @@ import {
   simpleRate,
   partialRate,
   generateOtherPerformanceMeasureData,
-} from "utils/testUtils/2023/validationHelpers";
+} from "utils/testUtils/validationHelpers";
 import { LabelData } from "utils";
 
 jest.mock("utils/getLabelText", () => ({
