+++ services/ui-src/src/measures/2023/shared/globalValidations/dataDrivenTools.test.ts
@@ -4,7 +4,7 @@ import {
   generateOmsQualifierRateData,
   generatePmQualifierRateData,
   simpleRate,
-} from "utils/testUtils/2023/validationHelpers";
+} from "utils/testUtils/validationHelpers";
 import {
   convertOmsDataToRateArray,
   getPerfMeasureRateArray,
