+++ services/ui-src/src/measures/2023/shared/globalValidations/validateEqualQualifierDenominators/index.test.ts
@@ -9,7 +9,7 @@ import {
   doubleRate,
   simpleRate,
   partialRate,
-} from "utils/testUtils/2023/validationHelpers";
+} from "utils/testUtils/validationHelpers";
 
 describe("Testing Equal Qualifier Denominators Across Category Validation", () => {
   const categories = [
