+++ services/ui-src/src/measures/2023/shared/globalValidations/validateNumeratorsLessThanDenominators/index.test.ts
@@ -10,7 +10,7 @@ import {
   simpleRate,
   partialRate,
   generateOtherPerformanceMeasureData,
-} from "utils/testUtils/2023/validationHelpers";
+} from "utils/testUtils/validationHelpers";
 
 describe("Testing Numerator Less Than Denominator", () => {
   const categories = [
