+++ services/ui-src/src/measures/2023/shared/globalValidations/validateOneCatRateHigherThanOtherCat/index.test.ts
@@ -10,7 +10,7 @@ import {
   higherRate,
   lowerRate,
   partialRate,
-} from "utils/testUtils/2023/validationHelpers";
+} from "utils/testUtils/validationHelpers";
 
 describe("Testing Category Rate Higher Than Other Validation", () => {
   const categories = [
