+++ services/ui-src/src/measures/2023/shared/globalValidations/validateTotals/index.test.tsx
@@ -1,6 +1,6 @@
 import { validateTotalNDR, validateOMSTotalNDR } from ".";
 
-import * as VH from "utils/testUtils/2023/validationHelpers";
+import * as VH from "utils/testUtils/validationHelpers";
 
 describe("Testing PM/OMS Total Validations", () => {
   describe("PM validation", () => {
