+++ services/ui-src/src/measures/2024/shared/globalValidations/validateEqualCategoryDenominators/index.test.ts
@@ -11,7 +11,7 @@ import {
   simpleRate,
   partialRate,
   generatePmQualifierRateData,
-} from "utils/testUtils/2024/validationHelpers";
+} from "utils/testUtils/validationHelpers";
 
 describe("Testing Equal Denominators For All Qualifiers Validation", () => {
   const noCat: LabelData[] = [];
