+++ services/ui-src/src/measures/2024/shared/globalValidations/validateRateZero/index.test.ts
@@ -10,7 +10,7 @@ import {
   simpleRate,
   partialRate,
   generateOtherPerformanceMeasureData,
-} from "utils/testUtils/2024/validationHelpers";
+} from "utils/testUtils/validationHelpers";
 
 describe("Testing Non-Zero/No Zero Numerator/Rate Validation", () => {
   const categories = [
