+++ services/ui-src/src/measures/2024/shared/globalValidations/validateSameDenominatorSets/index.test.ts
@@ -6,7 +6,7 @@ import {
   doubleRate,
   lowerRate,
   higherRate,
-} from "utils/testUtils/2024/validationHelpers";
+} from "utils/testUtils/validationHelpers";
 
 describe("Testing Same Denominator Set Validation", () => {
   const categories = [
