+++ services/ui-src/src/utils/testUtils/validationHelper.test.ts
@@ -9,15 +9,19 @@ const mockedConsoleError = jest.fn();
 describe("Test Validation Helpers", () => {
   it("should create valid rateData", () => {
     const data = VH.generateOmsQualifierRateData(
-      ["test1"],
-      ["test2"],
+      [{ label: "test1", id: "test1", text: "test1" }],
+      [{ label: "test2", id: "test2", text: "test2" }],
       [{ numerator: "5" }]
     );
-    expect(data.rates?.["test2"]?.["test1"]?.[0]?.numerator).toBe("5");
+    expect(data.rates?.["test1"]?.["test2"]?.[0]?.numerator).toBe("5");
   });
 
   it("should not create data if no data passed", () => {
-    const data = VH.generateOmsQualifierRateData(["test1"], ["test2"], []);
+    const data = VH.generateOmsQualifierRateData(
+      [{ label: "test1", id: "test1", text: "test1" }],
+      [{ label: "test2", id: "test2", text: "test2" }],
+      []
+    );
 
     expect(data).toEqual({});
     expect(mockedConsoleError).toHaveBeenCalled();
